added post route and refactored routes into conference.controller
