+++ src/routes/conference/conference.router.js
@@ -1,22 +1,13 @@
 const express = require('express');
 const router = express.Router();
-const {logger} = require('../../utils/logger');
-const {db} = require('../../api/db');
+const bodyParser = require("body-parser");
+const {getAllConferences, addNewConference} = require('../conference/conference.controller');
+const urlEncodedParser = bodyParser.urlencoded({extended: false});
 
-router.get('', (req, res) => {
-  let conferences = db.collection('conferences');
-  let events = [];
-  conferences.get().then((snapshot) => {
-    snapshot.forEach((doc) => {
-      events = [...events, doc.data()];
-    });
-    return res.send(events);
-  })
-  .catch((err) => {
-    logger.error(err);
-  });
-});
+router.get('', getAllConferences);
+
+router.post('', urlEncodedParser, addNewConference);
 
 module.exports = {
-  conferenceRouter: router
-}
+  conferenceRouter: router,
+};
